requirements.txt
requirements.txt:

fastapi
uvicorn
pydantic-settings
python-multipart
networkx
pypdf
sentence-transformers
faiss-cpu
openai
streamlit
httpx


...
